Run regresssion test with latest Gradle release only
