+++ subprojects/integ-test/src/crossVersionTest/groovy/org/gradle/integtests/TaskSubclassingBinaryCompatibilityCrossVersionSpec.groovy
@@ -43,7 +43,6 @@ import org.gradle.plugins.ide.idea.GenerateIdeaProject
 import org.gradle.plugins.ide.idea.GenerateIdeaWorkspace
 import org.gradle.plugins.signing.Sign
 import org.gradle.util.GradleVersion
-import spock.lang.Issue
 
 /**
  * Tests that task classes compiled against earlier versions of Gradle are still compatible.
@@ -210,106 +209,5 @@ apply plugin: SomePlugin
         then:
         version previous withTasks 'assemble' inDirectory(file("producer")) run()
         version current requireDaemon() requireIsolatedDaemons() withTasks 't' run()
-    }
-
-    @Issue("https://github.com/gradle/gradle/issues/16199")
-    def "can subclass task subclass in plugin"() {
-        setup:
-        file('plugin/settings.gradle') << """
-            rootProject.name = 'plugin'
-        """
-        file('plugin/build.gradle') << """
-            plugins {
-                id 'java-gradle-plugin'
-                id 'groovy'
-                id 'maven-publish'
-            }
-
-            group = "com.example"
-            version = "0.1.1"
-
-            repositories {
-                jcenter()
-            }
-
-            gradlePlugin {
-                plugins {
-                    sofPlugin {
-                        id = 'com.example.plugin'
-                        implementationClass = 'SofPlugin'
-                    }
-                }
-            }
-
-            publishing {
-                repositories {
-                    maven {
-                        name = "localRepo"
-                        setUrl(new File(project.buildDir, 'repo'))
-                    }
-                }
-            }
-        """
-        file("plugin/src/main/groovy/SofPlugin.groovy") << """
-            import org.gradle.api.Project
-            import org.gradle.api.Plugin
-
-            class SofPlugin implements Plugin<Project> {
-                void apply(Project project) {
-                    project.tasks.register('sofExec', CustomJavaExec) { task ->
-                        task.args = ["foo bar baz"] as List<String>
-                    }
-                }
-            }
-        """
-        file("plugin/src/main/groovy/CustomJavaExec.groovy") << """
-            import org.gradle.api.tasks.JavaExec
-
-            class CustomJavaExec extends CustomBaseJavaExec {
-
-                @Override
-                JavaExec setArgs(List<String> args) {
-                    println "args set: \$args"
-                    super.setArgs(args)
-                }
-            }
-        """
-        file("plugin/src/main/groovy/CustomBaseJavaExec.groovy") << """
-            import org.gradle.api.tasks.JavaExec
-
-            class CustomBaseJavaExec extends JavaExec {
-                // no setArgs overridden here
-            }
-        """
-        file("settings.gradle") << """
-            pluginManagement {
-                repositories {
-                    maven {
-                        url file('plugin/build/repo')
-                    }
-                }
-            }
-        """
-        file("build.gradle") << """
-            plugins {
-                id 'java-library'
-                id 'com.example.plugin' version '0.1.1'
-            }
-
-            tasks.named('sofExec') { JavaExec task ->
-                classpath = sourceSets.main.runtimeClasspath
-                main = "ClientMain"
-            }
-        """
-        file("src/main/java/ClientMain.java") << """
-            public class ClientMain {
-                public static void main(String[] args) {
-                }
-            }
-        """
-
-        expect:
-        version previous withTasks 'publish' inDirectory(file("plugin")) run()
-        version current requireDaemon() requireIsolatedDaemons() withTasks 'sofExec' run()
     }
 }
